% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{all_binary_sequences}
\alias{all_binary_sequences}
\title{Generate all possible sequences of 0s and 1s of a given length}
\usage{
all_binary_sequences(a)
}
\arguments{
\item{a}{the length of the sequences.}
}
\value{
a matrix containing in its rows the sequences of 0s and 1s.
}
\description{
Generate all possible sequences of 0s and 1s of a given length
}
