% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipaddgt2pop_misc.R
\name{corrUnif}
\alias{corrUnif}
\title{Generate correlated Uniform(0,1) random variables}
\usage{
corrUnif(n, Rho)
}
\arguments{
\item{n}{the desired number of sets of correlated Uniform random variables to generate}

\item{Rho}{the desired correlation matrix}
}
\value{
a matrix containing the realizations in the rows
}
\description{
Generate correlated Uniform(0,1) random variables
}
\examples{
n <- 1000
q <- 10
X <- corrUnif(n,Rho = (10/20)^abs(outer(1:q,1:q,"-")))

Rho
cor(X)
}
