% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd_misc.R
\name{dorfman.assay.gen}
\alias{dorfman.assay.gen}
\title{Generates Dorfman testing data.}
\usage{
dorfman.assay.gen(Y.true, Se, Sp, cj)
}
\arguments{
\item{Y.true}{The true disease statuses of the individuals.}

\item{Se}{A vector of testing sensitivities, where the first element is the
testing sensitivity for the master pools and the second entry is the
test sensitivity for individual testing.}

\item{Sp}{A vector of testing specificities, where the first element is the
testing specificity for the master pools and the second entry is the
test specificity for individual testing.}

\item{cj}{The size of the master pools (Note: The number of individuals \code{length(Y.true)} should be
evenly divisible by \code{cj}. This is only for decoding purposes; i.e.,
the regression methods do not require this condition).}
}
\value{
A list containing objects \code{Z} and \code{Y}.

This function simulates Dorfman decoding and stores the
testing responses in accordance to the data structure
required to fit the group testing regression model presented
in Gregory et al. (2018+). For the specifics of this structure
see McMahan et al. (2017).
}
\description{
Generates Dorfman testing data.
}
\examples{
Y.true <- rbinom(100,1,p=.05)
assay.data <- dorfman.assay.gen(Y.true,Se=c(.95,.92),Sp=c(.97,.98),cj=4)
}
