% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EYexact}
\alias{EYexact}
\title{Computes conditional expectations of individual disease statuses for individual, master pool, or Dorfman testing}
\usage{
EYexact(Z, Y, X, b, Se, Sp)
}
\arguments{
\item{Z}{Group testing output from one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}.}

\item{Y}{Group testing output from one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}.}

\item{X}{Design matrix with first column a column of 1s.}

\item{b}{Parameter values at which to compute the conditional expectations.}

\item{Se}{A vector of testing sensitivities of length \code{max(Z[,3])}.}

\item{Sp}{A vector of testing specificities of length \code{max(Z[,3])}.}
}
\value{
The vector of conditional expectations.

This function computes the conditional expectations of each individual disease status, conditional on the observed assay data and the diseasestatuses of all other individuals.
}
\description{
Computes conditional expectations of individual disease statuses for individual, master pool, or Dorfman testing
}
\examples{
grouplassogt2pop_data <- get_grouplassogt2pop_data( n1 = 400, n2 = 600)
  
EY <- EYexact(Z = grouplassogt2pop_data$Z1,
              Y = grouplassogt2pop_data$Y1,
              X = grouplassogt2pop_data$X1,
              b = rep(1,ncol(grouplassogt2pop_data$X1)),
              Se = grouplassogt2pop_data$Se1,
              Sp = grouplassogt2pop_data$Sp1)
}
