% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipaddgt2pop_misc.R
\name{get_grouplasso2pop_logreg_data}
\alias{get_grouplasso2pop_logreg_data}
\title{Generate two data sets with group testing responses and some common covariates}
\usage{
get_grouplasso2pop_logreg_data(n1, n2)
}
\arguments{
\item{n1}{the sample size for the first data set}

\item{n2}{the sample size for the second data set}
}
\value{
a list containing the data
}
\description{
Generate two data sets with group testing responses and some common covariates
}
