% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_grouplasso_data}
\alias{get_grouplasso_data}
\title{Generate a data set for group lasso}
\usage{
get_grouplasso_data(n, response)
}
\arguments{
\item{n}{the sample size}
}
\value{
a list containing the data
}
\description{
Generate a data set for group lasso
}
