% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd_misc.R
\name{get_grouplasso_logreg_data}
\alias{get_grouplasso_logreg_data}
\title{Generate a data set with binary responses}
\usage{
get_grouplasso_logreg_data(n)
}
\arguments{
\item{n}{the sample size}
}
\value{
a list containing the data
}
\description{
Generate a data set with binary responses
}
