% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_semipadd2pop_gt_data}
\alias{get_semipadd2pop_gt_data}
\title{Generate two data sets for semiparametric additive modeling with group testing responses and some common covariates}
\usage{
get_semipadd2pop_gt_data(n1, n2, model = 1)
}
\arguments{
\item{n1}{the sample size for the first data set}

\item{n2}{the sample size for the second data set}
}
\value{
a list containing the data
}
\description{
Generate two data sets for semiparametric additive modeling with group testing responses and some common covariates
}
