% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_semipadd_data}
\alias{get_semipadd_data}
\title{Generate a data set for semiparametric additive modeling with continuous responses that has an interaction effect}
\usage{
get_semipadd_data(n, response = "continuous")
}
\arguments{
\item{n}{the sample size}
}
\value{
a list containing the data
}
\description{
Generate a data set for semiparametric additive modeling with continuous responses that has an interaction effect
}
