% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd_misc.R
\name{get_semipadd_gt_data}
\alias{get_semipadd_gt_data}
\title{Generate a data set for semiparametric additive modeling with group testing responses and some common covariates}
\usage{
get_semipadd_gt_data(n)
}
\arguments{
\item{n}{the sample size}
}
\value{
a list containing the data
}
\description{
Generate a data set for semiparametric additive modeling with group testing responses and some common covariates
}
