% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd_misc.R
\name{get_semipadd_linreg_data}
\alias{get_semipadd_linreg_data}
\title{Generate a data set for semiparametric additive modeling with continuous responses}
\usage{
get_semipadd_linreg_data(n, extrafuns = 0)
}
\arguments{
\item{n}{the sample size}
}
\value{
a list containing the data
}
\description{
Generate a data set for semiparametric additive modeling with continuous responses
}
