% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipaddgt2pop_1pop.R
\name{grouplasso1pop_logreg_obj}
\alias{grouplasso1pop_logreg_obj}
\title{Compute the objective function of the 1-population group lasso problem with a binary response}
\usage{
grouplasso1pop_logreg_obj(beta1, Y1, X1, groups1, lambda, w1)
}
\arguments{
\item{beta1}{the vector of coefficients for data set 1}

\item{Y1}{the response vector of data set 1}

\item{X1}{matrix containing the design matrices for data set 1}

\item{groups1}{a vector indicating to which group each covariate of data set 1 belongs}

\item{lambda}{the level of sparsity penalization}

\item{w1}{group-specific weights for different penalization across groups in data set 1}
}
\value{
Returns the value of the objective function for the 2-population group lasso problem
}
\description{
Compute the objective function of the 1-population group lasso problem with a binary response
}
