% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd2pop_linreg.R
\name{grouplasso2pop_linreg_obj}
\alias{grouplasso2pop_linreg_obj}
\title{Compute the objective function of the 2-population group lasso problem with a continuous response}
\usage{
grouplasso2pop_linreg_obj(
  beta1,
  beta2,
  Y1,
  X1,
  groups1,
  Y2,
  X2,
  groups2,
  lambda,
  eta,
  w1,
  w2,
  w,
  AA1,
  AA2,
  Com
)
}
\arguments{
\item{beta1}{the vector of coefficients for data set 1}

\item{beta2}{the vector of coefficients for data set 2}

\item{Y1}{the response vector of data set 1}

\item{X1}{matrix containing the design matrices for data set 1}

\item{groups1}{a vector indicating to which group each covariate of data set 1 belongs}

\item{Y2}{the response vector of data set 2}

\item{X2}{matrix containing the design matrices for data set 2}

\item{groups2}{a vector indicating to which group each covariate of data set 2 belongs}

\item{lambda}{the level of sparsity penalization}

\item{eta}{the level of penalization towards model similarity}

\item{w1}{group-specific weights for different penalization across groups in data set 1}

\item{w2}{group-specific weights for different penalization across groups in data set 2}

\item{w}{group-specific weights for different penalization toward similarity for different groups}

\item{AA1}{a list of the matrices A2j}

\item{Com}{the indices of the covariate groups which are common in the two datasets}
}
\value{
Returns the value of the objective function for the 2-population group lasso problem
}
\description{
Compute the objective function of the 2-population group lasso problem with a continuous response
}
