% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouplasso2pop.R
\name{grouplasso2pop_linreg_R}
\alias{grouplasso2pop_linreg_R}
\title{Minimize the objective function of the 2-population group lasso problem with a continuous response}
\usage{
grouplasso2pop_linreg_R(
  Y1,
  X1,
  groups1,
  Y2,
  X2,
  groups2,
  rho1,
  rho2,
  lambda,
  eta,
  w1,
  w2,
  w,
  AA1,
  AA2,
  Com,
  tol = 1e-04,
  maxiter = 500,
  plot_obj = FALSE,
  init = NULL
)
}
\arguments{
\item{Y1}{the continuous response vector of data set 1}

\item{X1}{matrix containing the design matrices for data set 1}

\item{groups1}{a vector of integers indicating to which group each covariate in data set 1 belongs}

\item{Y2}{the binary response vector of data set 2}

\item{X2}{matrix containing the design matrices for data set 2}

\item{groups2}{a vector of integers indicating to which group each covariate in data set 1 belongs}

\item{rho1}{weight placed on the first data set}

\item{rho2}{weight placed on the second data set}

\item{lambda}{the level of sparsity penalization}

\item{eta}{the level of penalization towards model similarity}

\item{w1}{group-specific weights for different penalization across groups in data set 1}

\item{w2}{group-specific weights for different penalization across groups in data set 2}

\item{w}{group-specific weights for different penalization toward similarity for different groups}

\item{AA1}{a list of the matrices A2j}

\item{Com}{the indices of the covariate groups which are common in the two datasets}

\item{tol}{a convergence criterion}

\item{maxiter}{the maximum allowed number of iterations}

\item{init}{a list with vectors \code{beta1} and \code{beta2} to serve as initial values for the parameters}

\item{eigen1}{a list of eigen info on groups from data set 1}

\item{eigen2}{a list of eigen info on groups from data set 2}

\item{return_obj}{a logical indicating whether the value of the objection function should be recorded after every step of the algorithm}
}
\value{
Returns the minimizers of the 2-population group lasso objective function for the two data sets.
}
\description{
Minimize the objective function of the 2-population group lasso problem with a continuous response
}
\examples{
data <- get_grouplasso2pop_data(n1 = 400, n2 = 600, response = "continuous")

grouplasso2pop_linreg_R.out <- grouplasso2pop_linreg_R(Y1 = data$Y1,
                                                       X1 = data$X1,
                                                       groups1 = data$groups1,
                                                       Y2 = data$Y2,
                                                       X2 = data$X2,
                                                       groups2 = data$groups2,
                                                       rho1 = 2,
                                                       rho2 = 1,
                                                       lambda = 1,
                                                       eta = 1,
                                                       w1 = data$w1,
                                                       w2 = data$w2,
                                                       w = data$w,
                                                       AA1 = data$AA1,
                                                       AA2 = data$AA2,
                                                       Com = data$Com,
                                                       tol = 1e-4,
                                                       maxiter = 500)
}
