% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grouplasso2pop_logreg}
\alias{grouplasso2pop_logreg}
\title{Minimize the objective function of the 2-population group lasso problem with a binary response}
\usage{
grouplasso2pop_logreg(rY1, rX1, groups1, rY2, rX2, groups2, lambda, eta, w1, w2,
  w, rAA1, rAA2, rCom, tol, maxiter, beta1_init = as.numeric(c()),
  beta2_init = as.numeric(c()))
}
\arguments{
\item{groups1}{a vector of integers indicating to which group each covariate in data set 1 belongs}

\item{groups2}{a vector of integers indicating to which group each covariate in data set 1 belongs}

\item{lambda}{the level of sparsity penalization}

\item{eta}{the level of penalization towards model similarity}

\item{w1}{group-specific weights for different penalization across groups in data set 1}

\item{w2}{group-specific weights for different penalization across groups in data set 2}

\item{w}{group-specific weights for different penalization toward similarity for different groups}

\item{tol}{a convergence criterion}

\item{beta1_init}{optional starting value for beta1}

\item{beta2_init}{optional starting value for beta2}

\item{Y1}{the binary response vector of data set 1}

\item{X1}{matrix containing the design matrices for data set 1}

\item{Y2}{the binary response vector of data set 2}

\item{X2}{matrix containing the design matrices for data set 2}

\item{AA1}{a list of the matrices A1j}

\item{AA1}{a list of the matrices A2j}

\item{eigen1}{a list of eigen info on groups from data set 1}

\item{eigen2}{a list of eigen info on groups from data set 2}

\item{Com}{the indices of the covariate groups which are common in the two datasets}

\item{max.iter}{the maximum allowed number of iterations}

\item{return_obj}{a logical indicating whether the value of the objection function should be recorded after every step of the algorithm}
}
\value{
Returns the minimizers of the 2-population group lasso objective function for the two data sets.
}
\description{
Minimize the objective function of the 2-population group lasso problem with a binary response
}
