% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd2pop_logreg.R
\name{grouplasso2pop_logreg_cv}
\alias{grouplasso2pop_logreg_cv}
\title{Choose tuning parameters by crossvalidation for grouplasso2pop logreg.}
\usage{
grouplasso2pop_logreg_cv(
  Y1,
  X1,
  groups1,
  Y2,
  X2,
  groups2,
  rho1,
  rho2,
  n.lambda,
  n.eta,
  lambda.min.ratio,
  lambda.max.ratio = 1,
  n.folds,
  w1,
  w2,
  w,
  AA1,
  AA2,
  Com,
  tol = 1e-04,
  max.iter = 500,
  report.prog = TRUE
)
}
\arguments{
\item{Y1}{the binary response vector of data set 1}

\item{X1}{matrix containing the design matrices for data set 1}

\item{groups1}{a vector indicating to which group each covariate of data set 1 belongs}

\item{Y2}{the binary response vector of data set 2}

\item{X2}{matrix containing the design matrices for data set 2}

\item{groups2}{a vector indicating to which group each covariate of data set 2 belongs}

\item{rho1}{weight placed on the first data set}

\item{rho2}{weight placed on the second data set}

\item{n.lambda}{the number of lambda values desired}

\item{n.eta}{the number of eta values desired}

\item{lambda.min.ratio}{ratio of the smallest lambda value to the smallest value of lambda which admits no variables to the model}

\item{lambda.max.ratio}{ratio of the largest lambda value to the smallest value of lambda which admits no variables to the model}

\item{n.folds}{the number of crossvalidation folds}

\item{w1}{group-specific weights for different penalization across groups in data set 1}

\item{w2}{group-specific weights for different penalization across groups in data set 2}

\item{w}{group-specific weights for different penalization toward similarity for different groups}

\item{AA1}{a list of the matrices A2j}

\item{Com}{the indices of the covariate groups which are common in the two datasets}
}
\value{
a list containing the fits over a grid of lambda and eta values as well as the vector of lambda values and the vector of eta values
}
\description{
Choose tuning parameters by crossvalidation for grouplasso2pop logreg.
}
\examples{
grouplasso2pop_logreg_data <- get_grouplasso2pop_logreg_data(n1 = 400,
                                                             n2 = 600)

grouplasso2pop_logreg_cv.out <- grouplasso2pop_logreg_cv(Y1 = grouplasso2pop_logreg_data$Y1,
                                                         X1 = grouplasso2pop_logreg_data$X1,
                                                         groups1 = grouplasso2pop_logreg_data$groups1,
                                                         Y2 = grouplasso2pop_logreg_data$Y2,
                                                         X2 = grouplasso2pop_logreg_data$X2,
                                                         groups2 = grouplasso2pop_logreg_data$groups2,
                                                         rho1 = 2,
                                                         rho2 = 1,
                                                         n.lambda = 5,
                                                         n.eta = 5,
                                                         lambda.min.ratio = 0.01,
                                                         n.folds = 5,
                                                         w1 = grouplasso2pop_logreg_data$w1,
                                                         w2 = grouplasso2pop_logreg_data$w2,
                                                         w = grouplasso2pop_logreg_data$w,
                                                         AA1 = grouplasso2pop_logreg_data$AA1,
                                                         AA2 = grouplasso2pop_logreg_data$AA2,
                                                         Com = grouplasso2pop_logreg_data$Com,
                                                         tol = 1e-3,
                                                         max.iter = 500,
                                                         report.prog = TRUE)
}
