% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd2pop_logreg.R
\name{grouplasso2pop_logreg_R}
\alias{grouplasso2pop_logreg_R}
\title{Minimize the objective function of the 2-population group lasso problem with a binary response}
\usage{
grouplasso2pop_logreg_R(
  Y1,
  X1,
  groups1,
  Y2,
  X2,
  groups2,
  rho1,
  rho2,
  lambda,
  eta,
  w1,
  w2,
  w,
  AA1,
  AA2,
  Com,
  tol = 1e-04,
  max.iter = 500,
  plot_obj = FALSE,
  init = NULL
)
}
\arguments{
\item{Y1}{the binary response vector of data set 1}

\item{X1}{matrix containing the design matrices for data set 1}

\item{groups1}{a vector of integers indicating to which group each covariate in data set 1 belongs}

\item{Y2}{the binary response vector of data set 2}

\item{X2}{matrix containing the design matrices for data set 2}

\item{groups2}{a vector of integers indicating to which group each covariate in data set 1 belongs}

\item{rho1}{weight placed on the first data set}

\item{rho2}{weight placed on the second data set}

\item{lambda}{the level of sparsity penalization}

\item{eta}{the level of penalization towards model similarity}

\item{w1}{group-specific weights for different penalization across groups in data set 1}

\item{w2}{group-specific weights for different penalization across groups in data set 2}

\item{w}{group-specific weights for different penalization toward similarity for different groups}

\item{AA1}{a list of the matrices A2j}

\item{Com}{the indices of the covariate groups which are common in the two datasets}

\item{tol}{a convergence criterion}

\item{max.iter}{the maximum allowed number of iterations}

\item{init}{a list with vectors \code{beta1} and \code{beta2} to serve as initial values for the parameters}

\item{eigen1}{a list of eigen info on groups from data set 1}

\item{eigen2}{a list of eigen info on groups from data set 2}

\item{return_obj}{a logical indicating whether the value of the objection function should be recorded after every step of the algorithm}
}
\value{
Returns the minimizers of the 2-population group lasso objective function for the two data sets.
}
\description{
Minimize the objective function of the 2-population group lasso problem with a binary response
}
\examples{
# generate data
grouplasso2pop_logreg_data <- get_grouplasso2pop_logreg_data(n1 = 400,
                                                             n2 = 600)

# fit grouplasso2pop estimator
grouplasso2pop_logreg_R.out <- grouplasso2pop_logreg_R(Y1 = grouplasso2pop_logreg_data$Y1,
                                                   X1 = grouplasso2pop_logreg_data$X1,
                                                   groups1 = grouplasso2pop_logreg_data$groups1,
                                                   Y2 = grouplasso2pop_logreg_data$Y2,
                                                   X2 = grouplasso2pop_logreg_data$X2,
                                                   groups2 = grouplasso2pop_logreg_data$groups2,
                                                   rho1 = 2,
                                                   rho2 = 1,
                                                   lambda = 1,
                                                   eta = 1,
                                                   w1 = grouplasso2pop_logreg_data$w1,
                                                   w2 = grouplasso2pop_logreg_data$w2,
                                                   w = grouplasso2pop_logreg_data$w,
                                                   AA1 = grouplasso2pop_logreg_data$AA1,
                                                   AA2 = grouplasso2pop_logreg_data$AA2,
                                                   Com = grouplasso2pop_logreg_data$Com,
                                                   tol = 1e-4,
                                                   max.iter = 500,
                                                   plot_obj = TRUE)
}
