% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipaddgt2pop_misc.R
\name{grouplasso2pop_to_semipadd2pop}
\alias{grouplasso2pop_to_semipadd2pop}
\title{Convert output from grouplasso2pop to the fitted functions of the semi-parametric additive model}
\usage{
grouplasso2pop_to_semipadd2pop(X1, nonparm1, groups1, knots.list1, emp.cent1,
  QQ1.inv, b1, X2, nonparm2, groups2, knots.list2, emp.cent2, QQ2.inv, b2)
}
\arguments{
\item{X1}{the matrix with the observed covariate values for data set 1 (including a column of ones for the intercept)}

\item{nonparm1}{a vector indicating for which covariates a nonparametric function is to be estimated for data set 1}

\item{groups1}{a vector indicating to which group the entries of the coefficient vector \code{b1} belong}

\item{knots.list1}{a list of vectors with the knot locations for nonparametric effects for data set 1}

\item{emp.cent1}{a list of vectors of the empirical basis function centerings for data set 1}

\item{QQ1.inv}{the matrix with which to back-transform the group lasso coefficients for data set 1}

\item{b1}{the group lasso coefficients for data set 1}

\item{X2}{the matrix with the observed covariate values for data set 2 (including a column of ones for the intercept)}

\item{nonparm2}{a vector indicating for which covariates a nonparametric function is to be estimated for data set 2}

\item{groups2}{a vector indicating to which group the entries of the coefficient vector \code{b2} belong}

\item{knots.list2}{a list of vectors with the knot locations for nonparametric effects for data set 2}

\item{emp.cent2}{a list of vectors of the empirical basis function centerings for data set 2}

\item{QQ2.inv}{the matrix with which to back-transform the group lasso coefficients for data set 2}

\item{b2}{the group lasso coefficients for data set 2}
}
\description{
Convert output from grouplasso2pop to the fitted functions of the semi-parametric additive model
}
