% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouplasso.R
\name{grouplasso_gt_grid}
\alias{grouplasso_gt_grid}
\title{Fit grouplasso logistic regression estimator with group testing data over a grid of lambda values}
\usage{
grouplasso_gt_grid(
  Y,
  Z,
  Se,
  Sp,
  X,
  groups,
  n.lambda,
  lambda.min.ratio,
  lambda.max.ratio = 1,
  w,
  E.approx = FALSE,
  tol = 1e-04,
  maxiter = 500,
  report.prog = FALSE
)
}
\arguments{
\item{Y}{Group testing output in the format as output by one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}, or \code{array.assay.gen}.}

\item{Z}{Group testing output in the format as output by one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}, or \code{array.assay.gen}.}

\item{Se}{A vector of testing sensitivities, where the first element is the
testing specificity for pools and the second entry is the
test specificity for individual testing, if applicable.}

\item{Sp}{A vector of testing specificities, where the first element is the
testing specificity for pools and the second entry is the
test specificity for individual testing, if applicable.}

\item{X}{the matrix with the observed covariate values (including a column of ones for the intercept)}

\item{groups}{a vector indicating to which group each covariate belongs}

\item{n.lambda}{the number of lambda values desired}

\item{lambda.min.ratio}{ratio of the smallest lambda value to the smallest value of lambda which admits no variables to the model}

\item{w}{group-specific weights for different penalization of different groups}
}
\value{
a list containing the fits over a grid of lambda values as well as the vector of lambda values
}
\description{
Fit grouplasso logistic regression estimator with group testing data over a grid of lambda values
}
\examples{
grouplasso_gt_data <- get_grouplasso_gt_data(n = 1000)

grouplasso_gt_grid.out <- grouplasso_gt_grid(Y = grouplasso_gt_data$Y,
                                             Z = grouplasso_gt_data$Z,
                                             Se = grouplasso_gt_data$Se,
                                             Sp = grouplasso_gt_data$Sp,
                                             X = grouplasso_gt_data$X,
                                             groups = grouplasso_gt_data$groups,
                                             n.lambda = 10,
                                             lambda.min.ratio = 0.01,
                                             w  = grouplasso_gt_data$w,
                                             E.approx = FALSE,
                                             tol = 1e-3,
                                             maxiter = 500,
                                             report.prog = TRUE)
}
