% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grouplasso_linreg}
\alias{grouplasso_linreg}
\title{Minimize the objective function of the group lasso problem with a continuous response}
\usage{
grouplasso_linreg(
  rY,
  rX,
  groups,
  lambda,
  w,
  tol,
  maxiter,
  beta_init = as.numeric(c())
)
}
\arguments{
\item{groups}{a vector of integers indicating to which group each covariate belongs}

\item{lambda}{the level of sparsity penalization}

\item{w}{vector of group-specific weights for different penalization of groups}

\item{tol}{a convergence criterion}

\item{beta_init}{optional starting value for beta}

\item{Y}{the response vector}

\item{X}{matrix containing the design matrices}

\item{eigen}{a list of eigen info on groups}

\item{max.iter}{the maximum allowed number of iterations}

\item{return_obj}{a logical indicating whether the value of the objection function should be recorded after every step of the algorithm}
}
\value{
Returns the minimizer of the group lasso objective function
}
\description{
Minimize the objective function of the group lasso problem with a continuous response
}
\examples{
grouplasso_linreg_data <- get_grouplasso_linreg_data(n = 500)

grouplasso_linreg.out <- grouplasso_linreg(rY = grouplasso_linreg_data$Y,
                                           rX = grouplasso_linreg_data$X,
                                           groups = grouplasso_linreg_data$groups,
                                           lambda = 10,
                                           w = grouplasso_linreg_data$w,
                                           tol = 1e-4,                               
                                           maxiter = 500)
}
