% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd_linreg.R
\name{grouplasso_linreg_cv}
\alias{grouplasso_linreg_cv}
\title{Choose tuning parameters by crossvalidation for grouplasso linreg.}
\usage{
grouplasso_linreg_cv(
  Y,
  X,
  groups,
  n.lambda,
  lambda.min.ratio,
  lambda.max.ratio = 1,
  n.folds,
  w,
  tol = 1e-04,
  max.iter = 500,
  report.prog = TRUE
)
}
\arguments{
\item{Y}{the response vector}

\item{X}{matrix containing the design matrices}

\item{groups}{a vector indicating to which group each covariate belongs}

\item{n.lambda}{the number of lambda values desired}

\item{lambda.min.ratio}{ratio of the smallest lambda value to the smallest value of lambda which admits no variables to the model}

\item{n.folds}{the number of crossvalidation folds}

\item{w}{group-specific weights for different penalization toward similarity for different groups}
}
\value{
a list containing the fits over a grid of lambda values as well as the vector of lambda values
}
\description{
Choose tuning parameters by crossvalidation for grouplasso linreg.
}
\examples{
grouplasso_linreg_data <- get_grouplasso_linreg_data(n = 100)

grouplasso_linreg_cv.out <- grouplasso_linreg_cv(Y = grouplasso_linreg_data$Y,
                                                 X = grouplasso_linreg_data$X,
                                                 groups = grouplasso_linreg_data$groups,
                                                 n.lambda = 25,
                                                 lambda.min.ratio = 0.001,
                                                 lambda.max.ratio = 0.1,
                                                 n.folds = 5,
                                                 w = grouplasso_linreg_data$w,
                                                 tol = 1e-3,
                                                 max.iter = 500,
                                                 report.prog = FALSE)
}
