% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd_logreg.R
\name{grouplasso_logreg_cv}
\alias{grouplasso_logreg_cv}
\title{Choose tuning parameters by crossvalidation for grouplasso logreg.}
\usage{
grouplasso_logreg_cv(
  Y,
  X,
  groups,
  n.lambda,
  lambda.min.ratio,
  n.folds,
  w,
  tol = 1e-04,
  max.iter = 500,
  report.prog = TRUE
)
}
\arguments{
\item{Y}{the binary response vector}

\item{X}{matrix containing the design matrices}

\item{groups}{a vector indicating to which group each covariate belongs}

\item{n.lambda}{the number of lambda values desired}

\item{lambda.min.ratio}{ratio of the smallest lambda value to the smallest value of lambda which admits no variables to the model}

\item{n.folds}{the number of crossvalidation folds}

\item{w}{group-specific weights for different penalization toward similarity for different groups}
}
\value{
a list containing the fits over a grid of lambda values as well as the vector of lambda values
}
\description{
Choose tuning parameters by crossvalidation for grouplasso logreg.
}
\examples{
grouplasso_logreg_data <- get_grouplasso_logreg_data(n = 100)

grouplasso_logreg_cv.out <- grouplasso_logreg_cv(Y = grouplasso_logreg_data$Y,
                                                 X = grouplasso_logreg_data$X,
                                                 groups = grouplasso_logreg_data$groups,
                                                 n.lambda = 25,
                                                 lambda.min.ratio = 0.01,
                                                 n.folds = 5,
                                                 w = grouplasso_logreg_data$w,
                                                 tol = 1e-3,
                                                 max.iter = 500,
                                                 report.prog = FALSE)
}
