% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd_logreg.R
\name{grouplasso_logreg_cv_fixedgrid}
\alias{grouplasso_logreg_cv_fixedgrid}
\title{Choose tuning parameters by crossvalidation for grouplasso logreg when given a fixed grid of lambda values}
\usage{
grouplasso_logreg_cv_fixedgrid(
  Y,
  X,
  groups,
  lambda.seq,
  n.folds,
  b.init.mat,
  w,
  tol = 0.001,
  max.iter = 500
)
}
\arguments{
\item{Y}{the binary response vector}

\item{X}{matrix containing the design matrices}

\item{groups}{a vector indicating to which group each covariate belongs}

\item{lambda.seq}{sequence of lambda values}

\item{n.folds}{the number of crossvalidation folds}

\item{w}{group-specific weights for different penalization of different groups}

\item{tol}{the convergence tolerance}

\item{max.iter}{the maximum number of iterations allowed for each fit}

\item{b.init.arr}{array of initial values for beta}
}
\value{
a list containing the fits over a grid of lambda values as well as the vector of lambda values
}
\description{
Choose tuning parameters by crossvalidation for grouplasso logreg when given a fixed grid of lambda values
}
\examples{
grouplasso_logreg_data <- get_grouplasso_logreg_data(n = 100)

grouplasso_logreg_grid.out <- grouplasso_logreg_grid(Y = grouplasso_logreg_data$Y,
                                                     X = grouplasso_logreg_data$X,
                                                     groups = grouplasso_logreg_data$groups,
                                                     n.lambda = 25,
                                                     lambda.min.ratio = 0.01,
                                                     w = grouplasso_logreg_data$w,
                                                     tol = 1e-3,
                                                     max.iter = 500,
                                                     report.prog = FALSE)

lambda.seq <- grouplasso_logreg_grid.out$lambda.seq
b.mat <- grouplasso_logreg_grid.out$b.mat

grouplasso_logreg_cv_fixedgrid.out <- grouplasso_logreg_cv_fixedgrid(Y = grouplasso_logreg_data$Y,
                                                                     X = grouplasso_logreg_data$X,
                                                                     groups = grouplasso_logreg_data$groups,
                                                                     lambda.seq = lambda.seq,
                                                                     n.folds = 5,
                                                                     b.init.mat = b.mat,
                                                                     w = grouplasso_logreg_data$w,
                                                                     tol = 1e-3,
                                                                     max.iter = 500)
}
