% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouplasso.R
\name{grouplasso_logreg_grid}
\alias{grouplasso_logreg_grid}
\title{Fit grouplasso logistic regression estimator over a grid of lambda values}
\usage{
grouplasso_logreg_grid(
  Y,
  X,
  groups,
  n.lambda,
  lambda.min.ratio,
  lambda.max.ratio = 1,
  w,
  tol = 1e-04,
  maxiter = 500,
  report.prog = FALSE
)
}
\arguments{
\item{Y}{the binary response vector}

\item{X}{matrix containing the design matrices}

\item{groups}{a vector indicating to which group each covariate belongs}

\item{n.lambda}{the number of lambda values desired}

\item{lambda.min.ratio}{ratio of the smallest lambda value to the smallest value of lambda which admits no variables to the model}

\item{w}{group-specific weights for different penalization of different groups}
}
\value{
a list containing the fits over a grid of lambda values as well as the vector of lambda values
}
\description{
Fit grouplasso logistic regression estimator over a grid of lambda values
}
\examples{
grouplasso_logreg_data <- get_grouplasso_logreg_data(n = 400)

grouplasso_logreg_grid.out <- grouplasso_logreg_grid(Y = grouplasso_logreg_data$Y,
                                                     X = grouplasso_logreg_data$X,
                                                     groups = grouplasso_logreg_data$groups,
                                                     n.lambda = 25,
                                                     lambda.min.ratio = 0.001,
                                                     w = grouplasso_logreg_data$w,
                                                     tol = 1e-3,
                                                     maxiter = 500,
                                                     report.prog = TRUE)
}
