% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouplasso.R
\name{grouplasso_logreg_obj}
\alias{grouplasso_logreg_obj}
\title{Compute the objective function of the 1-population group lasso problem with a binary response}
\usage{
grouplasso_logreg_obj(beta, Y, X, groups, lambda, w)
}
\arguments{
\item{beta}{the vector of coefficients}

\item{Y}{the response vector}

\item{X}{matrix containing the design matrices}

\item{groups}{a vector indicating to which group each covariate of data set 1 belongs}

\item{lambda}{the level of sparsity penalization}

\item{w}{group-specific weights for different penalization of groups}
}
\value{
Returns the value of the objective function for the 2-population group lasso problem
}
\description{
Compute the objective function of the 1-population group lasso problem with a binary response
}
