% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd_logreg.R
\name{grouplasso_logreg_R}
\alias{grouplasso_logreg_R}
\title{Minimize the objective function of the 1-population group lasso problem with a binary response}
\usage{
grouplasso_logreg_R(
  Y,
  X,
  groups,
  lambda,
  w,
  tol = 1e-04,
  max.iter = 500,
  plot_obj = FALSE,
  init = NULL
)
}
\arguments{
\item{Y}{the binary response vector}

\item{X}{matrix containing the design matrices}

\item{groups}{a vector of integers indicating to which group each covariate belongs}

\item{lambda}{the level of sparsity penalization}

\item{tol}{a convergence criterion}

\item{max.iter}{the maximum allowed number of iterations}

\item{w1}{group-specific weights for different penalization across groups}

\item{return_obj}{a logical indicating whether the value of the objection function should be recorded after every step of the algorithm}
}
\value{
Returns the minimizer of the group lasso objective function
}
\description{
Minimize the objective function of the 1-population group lasso problem with a binary response
}
\examples{
# generate data
grouplasso_logreg_data <- get_grouplasso_logreg_data(n = 500)

# fit grouplasso1pop_logreg estimator
grouplasso_logreg.out <- grouplasso_logreg_R(Y = grouplasso_logreg_data$Y,
                                             X = grouplasso_logreg_data$X,
                                             groups = grouplasso_logreg_data$groups,
                                             lambda = 10,
                                             w = grouplasso_logreg_data$w,
                                             tol = 1e-4,
                                             max.iter = 500,
                                             plot_obj = TRUE)
}
