% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grouplasso_logreg_slower}
\alias{grouplasso_logreg_slower}
\title{Minimize the objective function of the group lasso problem with a binary response}
\usage{
grouplasso_logreg_slower(
  rY,
  rX,
  groups,
  lambda,
  w,
  tol,
  maxiter,
  beta_init = as.numeric(c())
)
}
\arguments{
\item{groups}{a vector of integers indicating to which group each covariate belongs}

\item{lambda}{the level of sparsity penalization}

\item{w}{vector of group-specific weights for different penalization of groups}

\item{tol}{a convergence criterion}

\item{beta_init}{optional starting value for beta}

\item{Y}{the binary response vector}

\item{X}{matrix containing the design matrices}

\item{eigen}{a list of eigen info on groups}

\item{max.iter}{the maximum allowed number of iterations}

\item{return_obj}{a logical indicating whether the value of the objection function should be recorded after every step of the algorithm}
}
\value{
Returns the minimizer of the group lasso objective function
}
\description{
Minimize the objective function of the group lasso problem with a binary response
}
\examples{
data <- get_grouplasso_data(n = 500, response = "binary")

grouplasso_logreg.out <- grouplasso_logreg(rY = data$Y,
                                           rX = data$X,
                                           groups = data$groups,
                                           lambda = 10,
                                           w = data$w,
                                           tol = 1e-4,
                                           maxiter = 500)
}
