% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{grouplasso_to_semipadd_noint}
\alias{grouplasso_to_semipadd_noint}
\title{Convert output from grouplasso to the fitted functions of the semi-parametric additive model}
\usage{
grouplasso_to_semipadd_noint(
  X,
  nonparm,
  groups,
  knots.list,
  emp.cent,
  QQ.inv,
  b
)
}
\arguments{
\item{X}{the matrix with the observed covariate values (including a column of ones for the intercept)}

\item{nonparm}{a vector indicating for which covariates a nonparametric function is to be estimated}

\item{groups}{a vector indicating to which group the entries of the coefficient vector \code{b} belong}

\item{knots.list}{a list of vectors with the knot locations for nonparametric effects}

\item{emp.cent}{a list of vectors of the empirical basis function centerings}

\item{QQ.inv}{the matrix with which to back-transform the group lasso coefficients}

\item{b}{the group lasso coefficients}
}
\description{
Convert output from grouplasso to the fitted functions of the semi-parametric additive model
}
