% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd2pop_gt.R
\name{grouplassogt2pop_cv_adapt_fixedgrid}
\alias{grouplassogt2pop_cv_adapt_fixedgrid}
\title{Choose tuning parameters for two population group lasso estimator with group testing data}
\usage{
grouplassogt2pop_cv_adapt_fixedgrid(
  Y1,
  Z1,
  Se1,
  Sp1,
  X1,
  groups1,
  Y2,
  Z2,
  Se2,
  Sp2,
  X2,
  groups2,
  rho1,
  rho2,
  lambda.seq,
  eta.seq,
  lambda.initial.fit,
  n.folds,
  w1,
  w2,
  w,
  AA1,
  AA2,
  Com,
  E.approx = FALSE,
  tol = 0.001,
  max.iter = 1000,
  report.prog = TRUE
)
}
\arguments{
\item{Y1}{Group testing output for data set 1 in the format as output by one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}, or \code{array.assay.gen}.}

\item{Z1}{Group testing output for data set 1 in the format as output by one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}, or \code{array.assay.gen}.}

\item{Se1}{A vector of testing sensitivities, where the first element is the
testing specificity for pools and the second entry is the
test specificity for individual testing, if applicable.}

\item{Sp1}{A vector of testing specificities, where the first element is the
testing specificity for pools and the second entry is the
test specificity for individual testing, if applicable.}

\item{X1}{the matrix with the observed covariate values for data set 1 (including a column of ones for the intercept)}

\item{groups1}{a vector indicating to which group each covariate of data set 2 belongs}

\item{Y2}{Group testing output for data set 2 in the format as output by one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}, or \code{array.assay.gen}.}

\item{Z2}{Group testing output for data set 2 in the format as output by one of the functions \code{individual.assay.gen}, \code{masterpool.assay.gen}, \code{dorfman.assay.gen}, or \code{array.assay.gen}.}

\item{Se2}{A vector of testing sensitivities, where the first element is the
testing specificity for pools and the second entry is the
test specificity for individual testing, if applicable.}

\item{Sp2}{A vector of testing specificities, where the first element is the
testing specificity for pools and the second entry is the
test specificity for individual testing, if applicable.}

\item{X2}{the matrix with the observed covariate values for data set 2 (including a column of ones for the intercept)}

\item{groups2}{a vector indicating to which group each covariate of data set 2 belongs}

\item{rho1}{weight placed on the first data set}

\item{rho2}{weight placed on the second data set}

\item{n.folds}{the number of crossvalidation folds}

\item{w1}{group-specific weights for different penalization across groups in data set 1}

\item{w2}{group-specific weights for different penalization across groups in data set 2}

\item{w}{group-specific weights for different penalization toward similarity for different groups}

\item{AA1}{a list of the matrices A2j}

\item{Com}{the indices of the covariate groups which are common in the two datasets}

\item{E.approx}{a logical indicating whether the conditional expectations in the E-step should be computed approximately or exactly.}

\item{tol}{a convergence criterion}

\item{max.iter}{the maximum allowed number of iterations (EM steps)}

\item{report.prog}{a logical. If \code{TRUE} then the number of inner loops required to complete the M step of the EM algorithm are returned after each EM step.}

\item{n.lambda}{the number of lambda values}

\item{n.eta}{the number of eta values}

\item{lambda.min.ratio}{ratio of the smallest lambda value to the smallest value of lambda which admits no variables to the model}
}
\value{
Returns the estimator of the parametric model with group testing data
}
\description{
Choose tuning parameters for two population group lasso estimator with group testing data
}
\examples{

}
