% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipaddgt2pop_misc.R
\name{individual.assay.gen}
\alias{individual.assay.gen}
\title{Generates individual testing data.}
\usage{
individual.assay.gen(Y.true, Se, Sp, cj = 1)
}
\arguments{
\item{Y.true}{The true disease statuses of the individuals.}

\item{Se}{The testing sensitivity used for individual testing.}

\item{Sp}{The testing specificity used for individual testing.}

\item{cj}{This is an extraneous argument and is here only in order that all four assay.gen functions take the same arguments. Default is \code{cj=1} and an error will be returns if \code{cj!=1}.}
}
\value{
a list containing objects \code{Z} and \code{Y}.

This function simulates individual level testing and stores the
testing responses in accordance to the data structure
required to fit the group testing regression model presented
in Gregory et al. (2018+). For the specifics of this structure
see McMahan et al. (2017).
}
\description{
Generates individual testing data.
}
\examples{
# generate individual covariate values and disease statuses
N <- 100
data <- model1(N)
X <- data$X
Y.true <- data$Yi
Se <- .95 # set assay sensitivity
Sp <- .97 # set assay specificity
# subject individuals to individual testing
assay.data <- individual.assay.gen(Y.true,Se,Sp,cj=1)
}
