% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{int.knots.at.empirical.quantiles}
\alias{int.knots.at.empirical.quantiles}
\title{Put knots at empirical quantiles, accommodating data with ties.}
\usage{
int.knots.at.empirical.quantiles(X, d)
}
\arguments{
\item{X}{the data}

\item{d}{the number of basis functions desired (cubic B-splines)}
}
\value{
the interior knots needed for constructing a centered, full-rank, cubic B-spline basis with \code{d} functions
}
\description{
Put knots at empirical quantiles, accommodating data with ties.
}
