% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{logit}
\alias{logit}
\title{The logit function}
\usage{
logit(z)
}
\arguments{
\item{z}{The value at which the logit function is to be evaluated.}
}
\value{
the value of the logit function
}
\description{
The logit function
}
