% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd_misc.R
\name{masterpool.assay.gen}
\alias{masterpool.assay.gen}
\title{Generates master pool testing data.}
\usage{
masterpool.assay.gen(Y.true, Se, Sp, cj)
}
\arguments{
\item{Y.true}{The true disease statuses of the individuals.}

\item{Se}{The master pool testing sensitivity.}

\item{Sp}{The master pool testing specificity.}

\item{cj}{The size of the master pools (Note: The number of individuals \code{length(Y.true)} should be
evenly divisible by \code{cj}, this is only for decoding purposes; i.e.,
the regression methods do not require this condition).}
}
\value{
a list containing objects \code{Z} and \code{Y}.

This function simulates Initial pool testing and stores the
testing responses in accordance to the data structure
required to fit the group testing regression model presented
in Gregory et al. (2018+). For the specifics of this structure
see McMahan et al. (2017).
}
\description{
Generates master pool testing data.
}
\examples{
# generate individual covariate values and disease statuses
N <- 100
data <- model1(N)
X <- data$X
Y.true <- data$Yi
Se <- .95 # set master pool assay sensitivity
Sp <- .97 # set master poot assay specificity
cj <- 4 # set size of master pools
# subject individuals to master pool testing
assay.data <- masterpool.assay.gen(Y.true,Se,Sp,cj)
}
