% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd2pop.R
\name{plot_semipadd2pop}
\alias{plot_semipadd2pop}
\title{A function for plotting the output of the function semipadd2pop}
\usage{
plot_semipadd2pop(x, true.functions = NULL)
}
\description{
A function for plotting the output of the function semipadd2pop
}
