% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot_semipadd2pop_cv}
\alias{plot_semipadd2pop_cv}
\title{Plot method for class semipadd2pop_gt_cv}
\usage{
plot_semipadd2pop_cv(x, true.functions = NULL)
}
\description{
Plot method for class semipadd2pop_gt_cv
}
