% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd2pop.R
\name{plot_semipadd2pop_cv_adapt}
\alias{plot_semipadd2pop_cv_adapt}
\title{A function for plotting the output of the function semipadd2pop_cv_adapt}
\usage{
plot_semipadd2pop_cv_adapt(x, true.functions = NULL)
}
\description{
A function for plotting the output of the function semipadd2pop_cv_adapt
}
