% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot_semipadd2pop_cv_wint}
\alias{plot_semipadd2pop_cv_wint}
\title{Plot method for class semipadd2pop_cv_wint}
\usage{
plot_semipadd2pop_cv_wint(x, true.functions = NULL)
}
\description{
Plot method for class semipadd2pop_cv_wint
}
