% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd_misc.R
\name{plot_semipadd2pop_gt}
\alias{plot_semipadd2pop_gt}
\title{Plot method for class semipadd2pop_gt}
\usage{
plot_semipadd2pop_gt(x, true.functions = NULL)
}
\description{
Plot method for class semipadd2pop_gt
}
