% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd_misc.R
\name{plot_semipadd2pop_gt_grid}
\alias{plot_semipadd2pop_gt_grid}
\title{Plot method for class semipadd2pop_gt_grid}
\usage{
plot_semipadd2pop_gt_grid(x, true.functions = NULL)
}
\description{
Plot method for class semipadd2pop_gt_grid
}
