% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd.R
\name{plot_semipadd_cv_adapt}
\alias{plot_semipadd_cv_adapt}
\title{Plot method for class semipadd_cv}
\usage{
plot_semipadd_cv_adapt(x, true.functions = NULL)
}
\description{
Plot method for class semipadd_cv
}
