% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipaddgt2pop_misc.R
\name{plot_semipaddgt2pop}
\alias{plot_semipaddgt2pop}
\title{Plot method for class semipaddgt}
\usage{
plot_semipaddgt2pop(x, true.functions = NULL)
}
\description{
Plot method for class semipaddgt
}
