% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipaddgt2pop_misc.R
\name{plot_semipaddgt2pop_cv}
\alias{plot_semipaddgt2pop_cv}
\title{Plot method for class semipaddgt_cv}
\usage{
plot_semipaddgt2pop_cv(x, true.functions = NULL)
}
\description{
Plot method for class semipaddgt_cv
}
