% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd_misc.R
\name{plot_semipaddgt2pop_grid}
\alias{plot_semipaddgt2pop_grid}
\title{Plot method for class semipaddgt2pop_grid}
\usage{
plot_semipaddgt2pop_grid(x, true.functions = NULL)
}
\description{
Plot method for class semipaddgt2pop_grid
}
