% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd_misc.R
\name{plot_semipaddgt_grid}
\alias{plot_semipaddgt_grid}
\title{Plot method for class semipaddgt_grid}
\usage{
plot_semipaddgt_grid(x, true.functions = NULL)
}
\description{
Plot method for class semipaddgt_grid
}
