% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pull.diagnoses}
\alias{pull.diagnoses}
\title{Pulls individual diagnoses from group testing data if available}
\usage{
pull.diagnoses(Z, Y)
}
\arguments{
\item{\code{Z}}{output from one of the functions \code{individial.assay.gen()},\code{masterpool.assay.gen()},\code{dorfman.assay.gen()}, or \code{array.assay.gen()}.}

\item{\code{Y}}{output from one of the functions \code{individial.assay.gen()},\code{masterpool.assay.gen()},\code{dorfman.assay.gen()}, or \code{array.assay.gen()}.}
}
\value{
a vector of 0s and 1s which are the individual diagnoses, NULL if \code{Z} and \code{Y} come from \code{masterpool.assay.gen()}.

This function pulls the individual diagnoses from
matrices Z and Y when individual diagnoses are available.
So for masterpool testing, NULL will be returned.
}
\description{
Pulls individual diagnoses from group testing data if available
}
