% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipaddgt2pop_misc.R
\name{semipadd2pop_to_grouplasso2pop}
\alias{semipadd2pop_to_grouplasso2pop}
\title{Prepare inputs for grouplasso2pop function when using it to fit a semiparametric model}
\usage{
semipadd2pop_to_grouplasso2pop(X1, nonparm1, X2, nonparm2, nCom, d1, d2, xi,
  w1 = 1, w2 = 1, w = 1, lambda.beta = 1, lambda.f = 1,
  eta.beta = 1, eta.f = 1)
}
\arguments{
\item{X1}{the matrix with the observed covariate values for data set 1 (including a column of ones for the intercept)}

\item{nonparm1}{a vector indicating for which covariates a nonparametric function is to be estimated for data set 1}

\item{X2}{the matrix with the observed covariate values for data set 2 (including a column of ones for the intercept)}

\item{nonparm2}{a vector indicating for which covariates a nonparametric function is to be estimated for data set 2}

\item{nCom}{the number of covariates to be treated as common between the two data sets: these must be arranged in the first \code{nCom} columns of the matrices \code{X1} and \code{X2} after the column of ones corresponding to the intercept.}

\item{d1}{vector giving the dimensions the  B-spline bases to be used when fitting the nonparametric effects. If a scalar is given, this dimension is used for all nonparametric effects.}

\item{d2}{vector giving the dimensions the  B-spline bases to be used when fitting the nonparametric effects. If a scalar is given, this dimension is used for all nonparametric effects.}

\item{xi}{a tuning parameter governing the smoothness of the nonparametric estimates}

\item{lambda.beta}{the level of sparsity penalization for the parametric effects}

\item{lambda.f}{the level of sparsity penalization for the nonparametric effects}

\item{eta.beta}{the level of penalization towards model similarity for parametric effects indicated to be common}

\item{eta.f}{the level of penalization towards model similarity for nonparametric effects indicated to be common}
}
\description{
Prepare inputs for grouplasso2pop function when using it to fit a semiparametric model
}
