% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd_linreg.R
\name{semipadd_linreg}
\alias{semipadd_linreg}
\title{Fit semiparametric regression model on continuous-response data}
\usage{
semipadd_linreg(
  Y,
  X,
  nonparm,
  w,
  d,
  xi,
  lambda.beta,
  lambda.f,
  tol = 1e-04,
  max.iter = 500,
  plot_obj = FALSE
)
}
\arguments{
\item{Y}{the response vector}

\item{X}{the matrix with the observed covariate values (including a column of ones for the intercept)}

\item{nonparm}{a vector indicating for which covariates a nonparametric function is to be estimated}

\item{w}{covariate-specific weights for different penalization toward similarity for different covariates}

\item{d}{vector giving the dimensions the  B-spline bases to be used when fitting the nonparametric effects. If a scalar is given, this dimension is used for all nonparametric effects.}

\item{xi}{a tuning parameter governing the smoothness of the nonparametric estimates}

\item{lambda.beta}{the level of sparsity penalization for the parametric effects}

\item{lambda.f}{the level of sparsity penalization for the nonparametric effects}

\item{tol}{a convergence criterion}

\item{max.iter}{the maximum allowed number of iterations}

\item{return_obj}{a logical indicating whether the value of the objection function should be recorded after every step of the algorithm}
}
\value{
Returns the estimator of the semiparametric additive model
}
\description{
Fit semiparametric regression model on continuous-response data
}
\examples{
semipadd_linreg_data <- get_semipadd_linreg_data(n = 200)

semipadd_linreg.out <- semipadd_linreg(Y = semipadd_linreg_data$Y,
                                       X = semipadd_linreg_data$X,
                                       nonparm = semipadd_linreg_data$nonparm,
                                       w = 1,
                                       d = 20,
                                       xi = 2,
                                       lambda.beta = 1,
                                       lambda.f = 1,
                                       tol = 1e-3,
                                       max.iter = 500,
                                       plot_obj = FALSE)

plot_semipaddgt(semipadd_linreg.out,
                true.functions = list(f = semipadd_linreg_data$f,
                                      X = semipadd_linreg_data$X)
                
)
}
