% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd.R
\name{semipadd_to_grouplasso}
\alias{semipadd_to_grouplasso}
\title{Prepare inputs for grouplasso function when using it to fit a semiparametric model}
\usage{
semipadd_to_grouplasso(
  X,
  nonparm,
  d,
  xi,
  w = 1,
  int = NULL,
  w_int = NULL,
  lambda.beta = 1,
  lambda.f = 1
)
}
\arguments{
\item{X}{the matrix with the observed covariate values (including a column of ones for the intercept)}

\item{nonparm}{a vector indicating for which covariates a nonparametric function is to be estimated}

\item{d}{vector giving the dimensions the  B-spline bases to be used when fitting the nonparametric effects. If a scalar is given, this dimension is used for all nonparametric effects.}

\item{xi}{a tuning parameter governing the smoothness of the nonparametric estimates}

\item{w}{covariate-specific weights for different penalization toward similarity for different covariates}

\item{int}{a matrix with rows giving the pairs of covariates for which to include interaction terms}

\item{w_int}{the penalization weights on the interaction terms}

\item{lambda.beta}{the level of sparsity penalization for the parametric effects}

\item{lambda.f}{the level of sparsity penalization for the nonparametric effects}
}
\description{
Prepare inputs for grouplasso function when using it to fit a semiparametric model
}
