% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipaddgt2pop_misc.R
\name{SoftThresh_R}
\alias{SoftThresh_R}
\title{The soft-thresholding function}
\usage{
SoftThresh_R(z, a)
}
\arguments{
\item{z}{the argument to which to apply soft-thresholding}

\item{a}{the threshold}
}
\value{
the value of the soft-thresholding function
}
\description{
The soft-thresholding function
}
\examples{
z <- seq(-4,4,length=21)
a <- 2
y <- SoftThresh(z,a)
plot(y~z,type="l")
}
