% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{spsm_cubespline_design}
\alias{spsm_cubespline_design}
\title{Get design matrix for finding regression splines coefficients under wiggliness penalization}
\usage{
spsm_cubespline_design(X, d, xi, W = NULL)
}
\arguments{
\item{X}{a vector of values at which penalized cubic B-splines are to be evaluated}

\item{d}{the number of B-spline functions in the basis.  If negative, knots are evenly spaced.}

\item{xi}{a parameter for penalizing towards smoothness}

\item{W}{a vector of observation weights}
}
\value{
a list with the knots, the empirical centerings of the columns, and some matrices
}
\description{
Get design matrix for finding regression splines coefficients under wiggliness penalization
}
