% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd2pop_linreg.R
\name{grouplasso2pop_linreg_cv_fixedgrid}
\alias{grouplasso2pop_linreg_cv_fixedgrid}
\title{Choose tuning parameters by crossvalidation for grouplasso2pop linreg when given a fixed grid of lambda and eta values}
\usage{
grouplasso2pop_linreg_cv_fixedgrid(
  Y1,
  X1,
  groups1,
  Y2,
  X2,
  groups2,
  lambda.seq,
  eta.seq,
  n.folds,
  b1.init.arr,
  b2.init.arr,
  w1,
  w2,
  w,
  AA1,
  AA2,
  Com,
  tol = 0.001,
  max.iter = 500
)
}
\arguments{
\item{Y1}{the continuous response vector of data set 1}

\item{X1}{matrix containing the design matrices for data set 1}

\item{groups1}{a vector indicating to which group each covariate of data set 1 belongs}

\item{Y2}{the continuous response vector of data set 2}

\item{X2}{matrix containing the design matrices for data set 2}

\item{groups2}{a vector indicating to which group each covariate of data set 2 belongs}

\item{lambda.seq}{sequence of lambda values}

\item{eta.seq}{sequence of eta values}

\item{n.folds}{the number of crossvalidation folds}

\item{b1.init.arr}{array of initial values for beta1}

\item{b2.init.arr}{array of initial values for beta2}

\item{w1}{group-specific weights for different penalization across groups in data set 1}

\item{w2}{group-specific weights for different penalization across groups in data set 2}

\item{w}{group-specific weights for different penalization toward similarity for different groups}

\item{AA1}{a list of the matrices A2j}

\item{Com}{the indices of the covariate groups which are common in the two datasets}

\item{tol}{the convergence tolerance}

\item{max.iter}{the maximum number of iterations allowed for each fit}
}
\value{
a list containing the fits over a grid of lambda and eta values as well as the vector of lambda values and the vector of eta values
}
\description{
Choose tuning parameters by crossvalidation for grouplasso2pop linreg when given a fixed grid of lambda and eta values
}
