% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipadd2pop_linreg.R
\name{grouplasso2pop_linreg_grid}
\alias{grouplasso2pop_linreg_grid}
\title{Fit grouplasso2pop regression estimator with continuous responses over a grid of lambda and eta values}
\usage{
grouplasso2pop_linreg_grid(
  Y1,
  X1,
  groups1,
  Y2,
  X2,
  groups2,
  n.lambda,
  n.eta,
  lambda.min.ratio,
  w1,
  w2,
  w,
  AA1,
  AA2,
  Com,
  tol = 1e-04,
  max.iter = 500,
  report.prog = FALSE,
  grid.only = FALSE
)
}
\arguments{
\item{Y1}{the continuous response vector of data set 1}

\item{X1}{matrix containing the design matrices for data set 1}

\item{groups1}{a vector indicating to which group each covariate of data set 1 belongs}

\item{Y2}{the continuous response vector of data set 2}

\item{X2}{matrix containing the design matrices for data set 2}

\item{groups2}{a vector indicating to which group each covariate of data set 2 belongs}

\item{n.lambda}{the number of lambda values desired}

\item{n.eta}{the number of eta values desired}

\item{lambda.min.ratio}{ratio of the smallest lambda value to the smallest value of lambda which admits no variables to the model}

\item{w1}{group-specific weights for different penalization across groups in data set 1}

\item{w2}{group-specific weights for different penalization across groups in data set 2}

\item{w}{group-specific weights for different penalization toward similarity for different groups}

\item{AA1}{a list of the matrices A2j}

\item{Com}{the indices of the covariate groups which are common in the two datasets}
}
\value{
a list containing the fits over a grid of lambda and eta values as well as the vector of lambda values and the vector of eta values
}
\description{
Fit grouplasso2pop regression estimator with continuous responses over a grid of lambda and eta values
}
